Revert "Merge branch 'master' of https://github.com/luo-liang/pytorch"
This reverts commit c030eaf7e647619ef8ef5548e42acf06b3ce1535, reversing
changes made to 0a0de7ec1bc14a3f6ffb1fd680815e7c8f5e35c0.

+++ torch/lib/c10d/ProcessGroupGloo.cpp
@@ -444,7 +444,7 @@ AlgorithmEntry* ProcessGroupGloo::checkout(const AlgorithmKey& key) {
   const auto i = cacheCurrentEntry_[key];
 
   // Ensure the cache vector is appropriately sized
-  if (vec.size() != static_cast<size_t>(cacheNumAlgorithmEntries_)) {
+  if (vec.size() != cacheNumAlgorithmEntries_) {
     vec.resize(cacheNumAlgorithmEntries_);
   }
